dEngine 2024
It is still in development, but I'd rather push it as it is now rather than wait and wait.
Maybe I will make a wiki, maybe I won't, who knows.

This is NOT a gamemode. It's a framework that's supposed to HELP with making a gamemode.

Every change that shouldn't be considered default/vanilla (eg. stamina, professions) or features that block the customization (eg. forced loot tables) will NEVER be in this repository. It can be in an eventual fork or a gamemode though.